Using Numba to JIT compile compute-heavy functions
After having shifted around the structure and logic of the system for
hours, I felt like having an easy win. So, I used Numba to JIT compile
the compute-intensive functions, and (unsurprisingly) it massively
improved the force matrix calculations. The improvement to boundary
evaluation is debatable, but it should scale a little better.