Exhaustive state saving
If `save = True`, a directory constructed and populated with a
`params.json` file and a number of compressed .npz files for each
iteration based on the desired frequency. I have opted for a space
inefficent approach so that simulation loading is non-resource
intensive after the fact. To load a given iteration; load the file
and draw the plots - no computation required.