Beginnings of animation and fixing connectivity
This commit has the rudimentary beginnings of the matplotlib based
animation functionality - there is still a long, long way to go on
this. It should serve as a good structure on which to base things
though. This does not rule out interactive plots, I quite liked
those. I also found that adjacencies were being duplicated, so I
fixed that.